+++ src/main/java/com/codessquad/qna/users/UserService.java
@@ -2,16 +2,44 @@ package com.codessquad.qna.users;
 
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 @Service
-public interface UserService {
-    List<User> getUsers();
+public class UserService {
+    private final List<User> users = Collections.synchronizedList(new ArrayList<>());
 
-    void addUser(User user);
+    public List<User> getUsers() {
+        return users;
+    }
 
-    void updateUser(User user);
+    public void addUser(User newUser) {
+        for (User user : users) {
+            if (user.getUserId().equals(newUser.getUserId())) {
+                return;
+            }
+        }
 
-    Optional<User> getUser(String userId);
+        users.add(newUser);
+    }
+
+    public void updateUser(User toUpdate) {
+        for (User user : users) {
+            if (user.equals(toUpdate)) {
+                user.update(toUpdate);
+                return;
+            }
+        }
+    }
+
+    public Optional<User> getUser(String userId) {
+        for (User user : users) {
+            if (user.getUserId().equals(userId)) {
+                return Optional.of(user);
+            }
+        }
+        return Optional.empty();
+    }
 }
